// src/modules/Invoice/usecase/find-by-id/find-by-id.dto.ts
export interface FindByIdInvoiceUseCaseInputDto {
    id: string;
}

export interface FindByIdInvoiceUseCaseOutputDto {
    id: string;
    name: string;
    document: string;
    address: {
        street: string;
        number: string;
        complement: string;
        city: string;
        state: string;
        zipCode: string;
    };
    items: {
        id: string;
        name: string;
        price: number;
    }[];
    total: number;
    createdAt: Date;
}